requirements.txt
faiss-cpu==1.10.0
fastapi==0.115.12
hydra-core==1.3.2
numpy==1.26.4
omegaconf==2.3.0
pandas==1.5.0
python-dotenv==1.0.1
streamlit==1.48.1
tokenizers==0.21.1
torch==2.6.0
transformers==4.51.3
uvicorn==0.34.3
vllm==0.8.5
